+++ lib/brando/villain.ex
@@ -339,16 +339,19 @@ defmodule Brando.Villain do
   def map_images(images) do
     Enum.map(images, fn image_record ->
       img_field = image_record.image
+
       sizes =
         img_field.sizes
         |> Enum.map(&({elem(&1, 0), media_url(elem(&1, 1))}))
         |> Enum.into(%{})
 
-      %{src: media_url(img_field.path),
-        thumb: media_url(img_url(img_field, :thumb)),
-        sizes: sizes,
-        title: img_field.title,
-        credits: img_field.credits}
+      %{
+        src:     media_url(img_field.path),
+        thumb:   media_url(img_url(img_field, :thumb)),
+        sizes:   sizes,
+        title:   img_field.title,
+        credits: img_field.credits
+      }
     end)
   end
 end